memperbaiki tbl migration

+++ database/migrations/2022_12_05_050548_tbl_bahan_masuk.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('kd_bahan', 10);
             $table->string('nm_bahan', 50);
             $table->date('tgl_masuk');
-            $table->float('jumlah');
+            $table->double('jumlah', 10);
             $table->string('ket');
             $table->double('total');
             $table->timestamps();
