+++ database/migrations/2022_12_05_065949_tbl_data_bahan.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('kd_bahan', 10)->primary();
             $table->string('nm_bahan', 50);
             $table->integer('harga_beli');
-            $table->float('stok');
+            $table->double('stok');
             $table->string('ket');
             $table->timestamps();
         });
