+++ database/migrations/2022_12_06_034921_tbl_produk_jadi.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->integer('stok');
             $table->integer('modal');
             $table->integer('harga_jual');
-            $table->float('berat');
+            $table->double('berat', 10);
             $table->string('ket');
             $table->text('foto');
